Split out separate parsing file
lib.rs is getting huge and unwieldy.

This is a pure code move.
